// src/data/ALL_EXPERIENCES.d.ts
export declare const ALL_EXPERIENCES: {
    job1: {
        title: string;
        employer: string;
        period: string;
        type: string;
        details: string[];
    };
    job2: {
        title: string;
        employer: string;
        period: string;
        type: string;
        details: string[];
    };
    job3: {
        title: string;
        employer: string;
        period: string;
        type: string;
        details: string[];
    };
    job4: {
        title: string;
        employer: string;
        period: string;
        type: string;
        details: string[];
    };
    volunteer1: {
        title: string;
        employer: string;
        period: string;
        type: string;
        details: string[];
    };
};
//# sourceMappingURL=ALL_EXPERIENCES.d.ts.map